+++ NEWS.md
@@ -11,6 +11,10 @@
 * Added `more_than` option to `calc_n_samples` so that users can set the minimum threshold for whether a sample is counted or not instead of it always 1.
 * Added `calc_prop_samples` function for calculating the proportion of samples with a value greater than 0 (issues [#233](https://github.com/grunwaldlab/metacoder/issues/233).
 
+### New features
+
+* `taxmap` objects can be converted to `phyloseq` objects using `as_phyloseq`.
+
 ##  metacoder 0.2.1
 
 ### Bug fixes